+++ _pytest/junitxml.py
@@ -237,7 +237,13 @@ def record_property(request):
     """Add an extra properties the calling test.
     User properties become part of the test report and are available to the
     configured reporters, like JUnit XML.
-    The fixture is callable with ``(name, value)``.
+    The fixture is callable with ``(name, value)``, with value being automatically
+    xml-encoded.
+
+    Example::
+
+        def test_function(record_property):
+            record_property("example_key", 1)
     """
     request.node.warn(
         code='C3',